# Enable `Performance/UnfreezeString` cop