--- a/actionview/lib/action_view/helpers/translation_helper.rb
+++ b/actionview/lib/action_view/helpers/translation_helper.rb
@@ -98,7 +98,7 @@ module ActionView
           raise e if raise_error
 
           keys = I18n.normalize_keys(e.locale, e.key, e.options[:scope])
-          title = "translation missing: #{keys.join('.')}".dup
+          title = +"translation missing: #{keys.join('.')}"
 
           interpolations = options.except(:default, :scope)
           if interpolations.any?