--- a/actionview/lib/action_view/log_subscriber.rb
+++ b/actionview/lib/action_view/log_subscriber.rb
@@ -16,7 +16,7 @@ module ActionView
 
     def render_template(event)
       info do
-        message = "  Rendered #{from_rails_root(event.payload[:identifier])}".dup
+        message = +"  Rendered #{from_rails_root(event.payload[:identifier])}"
         message << " within #{from_rails_root(event.payload[:layout])}" if event.payload[:layout]
         message << " (#{event.duration.round(1)}ms)"
       end
@@ -24,7 +24,7 @@ module ActionView
 
     def render_partial(event)
       info do
-        message = "  Rendered #{from_rails_root(event.payload[:identifier])}".dup
+        message = +"  Rendered #{from_rails_root(event.payload[:identifier])}"
         message << " within #{from_rails_root(event.payload[:layout])}" if event.payload[:layout]
         message << " (#{event.duration.round(1)}ms)"
         message << " #{cache_message(event.payload)}" unless event.payload[:cache_hit].nil?
@@ -85,7 +85,7 @@ module ActionView
 
     def log_rendering_start(payload)
       info do
-        message = "  Rendering #{from_rails_root(payload[:identifier])}".dup
+        message = +"  Rendering #{from_rails_root(payload[:identifier])}"
         message << " within #{from_rails_root(payload[:layout])}" if payload[:layout]
         message
       end