--- a/activerecord/lib/arel/collectors/plain_string.rb
+++ b/activerecord/lib/arel/collectors/plain_string.rb
@@ -4,7 +4,7 @@ module Arel # :nodoc: all
   module Collectors
     class PlainString
       def initialize
-        @str = "".dup
+        @str = +""
       end
 
       def value