@matthewd 2018-05-24
Extra thought: is there a cop we can use in 5-2-stable to disallow this syntax? Accidentally backporting a missing `.freeze` from #32031 would [only] harm performance on 2.2, whereas an accidental backport here would be an error (though hopefully caught by the test suite).