@y-yagi 2018-05-24
I agree that this will be an obstacle to the backport. 
So, for a while, I'd like to wait to merge this until 5.2.1 is released. I think that it is sufficient if 5.2.1 is released. What about it? 

> is there a cop we can use in 5-2-stable to disallow this syntax? 

Unfortunately, there seems to be no cop for that. 
We added a test that was missing when dealing with `dup`, so I think that it can probably be checked by testing. But I understand that there is concern. Add a custom cop?